// angular-sample/angular-webapp/client/app/risk-assessment/models/risk-assessment-control.ts
export class RiskAssessmentControl {
    Id: string;
    Name: string;
    CompanyId: string;
    Description: string;
    PictureId: string;
    RiskAssessmentId: string;
    PrototypeId: string;
    IsSharedPrototype: boolean;
    RiskAssessmentHazardId: string;
    Category: number;
    IsExample: boolean;
}